[解决编译问题]

+++ credis/credis/cconfig/ccfg.cpp
@@ -273,7 +273,7 @@ namespace chen {
 		{
 			return;
 		}
-		int i, start_pos, end_pos;
+		size_t i, start_pos, end_pos;
 		for (i = 0; i < str.size(); i++)
 		{
 			if (!_is_space(str[i]))
